+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 [unreleased] - 2022-11-29
+### Added
+- Ambient input from ROMS ocean model
+
 ### Changed
 - Use proper dates in input/output instead of seconds since simulation start